# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Answering Service Portal API
  version: "1.0.0"
  description: |
    The Answering Service Portal REST API for operators and their integrators.

    All endpoints require a Bearer token in the `Authorization` header.
    API keys are created in the Operator Admin under **Settings → API Keys**.

    Keys are either **operator-scoped** (access to all clients in your org) or
    **business-scoped** (access to a single client). Several write endpoints
    require operator-scoped keys.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://your-portal.com
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        API key created in Operator Admin → Settings → API Keys.
        Format: `Authorization: Bearer sk_live_...`
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, code]
          properties:
            message:
              type: string
              example: "business_id query parameter is required for operator keys."
            code:
              type: string
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - BAD_REQUEST
                - NOT_FOUND
                - INTERNAL_ERROR
              example: BAD_REQUEST
    CallInput:
      type: object
      required: [businessId, timestamp, callType, direction, durationSeconds, telephonyStatus, message]
      properties:
        businessId:
          type: string
          format: uuid
          description: UUID of the client business. Must belong to your operator org.
        timestamp:
          type: string
          format: date-time
          description: When the call occurred (ISO 8601).
        callerName:
          type: string
          description: Caller's full name.
        callerNumber:
          type: string
          description: Caller's phone number.
        callbackNumber:
          type: string
          description: Callback number (may differ from callerNumber).
        callType:
          type: string
          enum: [urgent, new-client, appointment, general-info, after-hours]
          description: Call type slug.
        direction:
          type: string
          enum: [inbound, outbound]
        durationSeconds:
          type: integer
          minimum: 0
          description: Duration in seconds. Use 0 for missed/voicemail.
        telephonyStatus:
          type: string
          enum: [completed, missed, voicemail]
        message:
          type: string
          minLength: 1
          description: Agent's written call note. Must not be empty.
        recordingUrl:
          type: string
          format: uri
          description: URL to a voicemail recording.
    CallIngestResult:
      type: object
      required: [businessId, status]
      properties:
        businessId:
          type: string
          format: uuid
        status:
          type: string
          enum: [inserted, error]
        callId:
          type: string
          format: uuid
          description: Present when status is "inserted".
        issue:
          type: string
          description: Validation error message. Present when status is "error".
...
    CallLogSummary:
      type: object
      description: >
        Call log record as returned by the list endpoint. Does not include caller contact fields
        (caller_name, caller_number, callback_number, has_recording) — use GET /api/v1/calls/{id}
        for the full detail.
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        call_type:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        duration_seconds:
          type: integer
        telephony_status:
          type: string
          enum: [completed, missed, voicemail]
        message:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        portal_status:
          type: string
          enum: [new, read, flagged_qa, assigned, resolved]
...
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
...
paths:
  /api/v1/calls:
    get:
      summary: List call logs
      description: >
        Returns paginated call logs for a business.
        Operator keys require the `business_id` query parameter.
      operationId: listCalls
      tags: [Calls]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (1-indexed).
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
          description: Results per page.
        - name: business_id
          in: query
          schema:
            type: string
            format: uuid
          description: Required for operator-scoped keys.
      responses:
        "200":
          description: Paginated call log list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CallLogSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "400":
          description: Missing required parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Invalid or missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Ingest calls
      description: >
        Ingest one or more call records. Accepts a JSON array or a CSV file
        (`multipart/form-data` with a `file` field).
        Requires an **operator-scoped** key with `calls:write` scope.
      operationId: ingestCalls
      tags: [Calls]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CallInput'
              minItems: 1
            example:
              - businessId: "4a3b2c1d-0000-0000-0000-000000000001"
                timestamp: "2026-03-26T14:30:00Z"
                callerName: "John Smith"
                callerNumber: "+15555550100"
                callbackNumber: "+15555550100"
                callType: "urgent"
                direction: "inbound"
                durationSeconds: 180
                telephonyStatus: "completed"
                message: "Caller reports chest pain. Requesting callback from Dr. Jones."
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with call records.
      responses:
        "201":
          description: All rows inserted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      inserted:
                        type: integer
                      errors:
                        type: integer
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/CallIngestResult'
        "207":
          description: Partial success — some rows failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      inserted:
                        type: integer
                      errors:
                        type: integer
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/CallIngestResult'
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Invalid or missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Business-scoped key used; operator key required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...